refactor: move prefs_storage inline tests to sibling tests.rs
Shrinks the ~123-line inline #[cfg(test)] mod tests block out of
prefs_storage.rs into a sibling prefs_storage/tests.rs, following the
db/src/books/tests.rs pattern (#1267). Pure file move, no logic change.
